Conflicts:
	htdocs/reporter/db/stats/last-updated.json
	htdocs/reporter/db/stats/views-2012-09.json
	htdocs/stat/browsers-plain/data.xml
	htdocs/stat/browsers/data.xml
	htdocs/stat/cities/data.xml
	htdocs/stat/errors/data.xml
	htdocs/stat/visitors/data.xml

+++ barman/processors/analytics.rb
@@ -3,12 +3,15 @@
 
 require "digest/md5"
 require "optparse"
+require "date"
 
 require "lib/json"
 require "lib/string"
 require "lib/array"
 require "lib/file"
 require "lib/output"
+require "lib/curl"
+require "lib/oauth2helper"
 
 require "config"
 require "entities/entity"
@@ -20,29 +23,30 @@ class Analytics
   HOUR = MINUTE * 60
   DAY  = 24 * 60 * 60
   
-  
   module Config
     PROFILE_ID     = "9038802"
-    BASE_DIR       = Inshaker::BASE_DIR + "Blog/"
-    CLIENT_ID      = "3164701909-cl0sa37gnh889cr5f043t6aeim88r7gk.apps.googleusercontent.com"
-    TOKEN_URI      = "https://accounts.google.com/o/oauth2/token"
     DATA_URI       = "https://www.googleapis.com/analytics/v3/data/ga"
-    SECRET         = ENV["ANALYTICS_CLIENT_SECRET"]
-    TOKEN_REFRESH  = "1/db0zlC0q9jiRo6vlQ45zWnFx32ER3orsVS089-NKCao"
     
     TMP            = Inshaker::ROOT_DIR + "/barman/tmp"
-    TOKEN_ACCESS   = TMP + "/token-access.txt"
     
+    REPORTER_DIR   = Inshaker::HTDOCS_DIR + "/reporter/db/stats"
+    ALL_JSON       = REPORTER_DIR + "/all.json"
+    LAST_UP_JSON   = REPORTER_DIR + "/last-updated.json"
     
-    HT_STAT_DIR    = Inshaker::HTDOCS_DIR + "/reporter/db/stats"
-    ALL_JSON       = HT_STAT_DIR + "/all.json"
-    LAST_UP_JSON   = HT_STAT_DIR + "/last-updated.json"
+    HT_STAT_DIR    = Inshaker::HTDOCS_DIR + "/db/stats"
+    VISITS_JSON    = HT_STAT_DIR + "/visits.json"
+    CITIES_JSON    = HT_STAT_DIR + "/cities.json"
+    BROWSERS_JSON  = HT_STAT_DIR + "/browsers.json"
+    BROWSERSP_JSON = HT_STAT_DIR + "/browsers-plain.json"
+    ERRORS_JSON    = HT_STAT_DIR + "/errors.json"
     
     HT_RATING_JSON = Inshaker::HTDOCS_DIR + "/db/ratings/rating.json"
   end
   
   def initialize
     @all = []
+    @auth = OAuth2Helper.new
+    @auth.temp = Config::TMP
   end
   
   def process_options
@@ -52,6 +56,10 @@ class Analytics
       
       opts.on("-q", "--quite", "сообщать только об ошибках") do |v|
         quite!
+      end
+      
+      opts.on("-f", "--force", "не кешировать ответы") do |v|
+        @options[:force] = true
       end
     end.parse!
   end
@@ -72,56 +80,9 @@ class Analytics
   end
   
   def get_credentials
-    
-    check_auth and return true
-    
-    refresh and check_auth and return true
-    
-    # login and check_auth and return true
-    
-    return false
+    @token = @auth.get_access_token
   end
   
-  def ping
-    r = raw_get(Config::DATA_URI + "?ids=ga:#{Config::PROFILE_ID}&dimensions=ga:pagePath&metrics=ga:pageviews&start-date=2010-04-20&end-date=2010-05-20&max-results=10")
-    puts r
-    JSON.parse(r)["kind"] == "analytics#gaData"
-  end
-  
-  def check_auth
-    
-    unless @token
-      unless File.exists?(Config::TOKEN_ACCESS)
-        return false
-      end
-      
-      unless Time.now - File.mtime(Config::TOKEN_ACCESS) < HOUR
-        return false
-      end
-      
-      @token = File.read(Config::TOKEN_ACCESS)
-    end
-    
-    ping
-  end
-  
-  def refresh
-    
-    r = IO.popen(["curl", "-s", "-d", "client_id=#{Config::CLIENT_ID}", "-d", "client_secret=#{Config::SECRET}", "-d", "refresh_token=#{Config::TOKEN_REFRESH}", "-d", "grant_type=refresh_token", Config::TOKEN_URI]).read
-    
-    # puts r
-    
-    r = JSON.parse(r)
-    
-    unless r["access_token"]
-      return false
-    end
-    
-    @token = r["access_token"]
-    File.write(Config::TOKEN_ACCESS, @token)
-    
-    true
-  end
   
   def get_last_updated
     @last_updated = Time.at(JSON.parse(File.read(Config::LAST_UP_JSON))[0])
@@ -135,22 +96,22 @@ class Analytics
     File.exists?(fn) && Time.now - File.mtime(fn) < sec
   end
   
-  def raw_get url
-    io = IO.popen(["curl", "-s", "-H", "Authorization: Bearer #{@token}", url])
-    r = io.read
-    io.close
-    
-    return r
+  def get_authed url, query
+    Curl.get(url, query, {"Authorization" => "Bearer #{@token}"})
   end
   
-  def get url
-    hash = Digest::MD5.hexdigest(url)
+  def get_cached url, query
+    if @options[:force]
+      return get_authed(url, query)
+    end
+    
+    hash = Digest::MD5.hexdigest("#{url}?#{query.to_a.flatten.join("&")}")
     cache = "#{Config::TMP}/#{hash}.url.txt"
     if newer?(cache, 15 * MINUTE)
       return File.read(cache)
     end
     
-    r = raw_get(url)
+    r = get_authed(url, query)
     
     File.write(cache, r)
     
@@ -158,20 +119,31 @@ class Analytics
   end
   
   def report query, start, endd, results=100
-    get Config::DATA_URI +
-        "?ids=ga:#{Config::PROFILE_ID}" +
-        "&#{query}&start-date=#{start.strftime("%Y-%m-%d")}&end-date=#{endd.strftime("%Y-%m-%d")}&max-results=#{results}&prettyprint=true"
+    rep =
+    {
+      "ids" => "ga:#{Config::PROFILE_ID}",
+      "start-date" => start.strftime("%Y-%m-%d"),
+      "end-date" => endd.strftime("%Y-%m-%d"),
+      "max-results" => results
+    }
+    get_cached(Config::DATA_URI, query.merge(rep))
   end
   
   def get_pageviews start, endd
-    json = report("dimensions=ga:pagePath&metrics=ga:pageviews,ga:uniquePageviews&filters=ga:pagePath=~^/cocktails?/&sort=-ga:pageviews", start, endd, 10000)
+    query =
+    {
+      "dimensions" => "ga:pagePath",
+      "metrics" => "ga:pageviews,ga:uniquePageviews",
+      "filters" => "ga:pagePath=~^/cocktails?/",
+      "sort" => "-ga:pageviews"
+    }
+    json = report(query, start, endd, 10000)
     data = JSON.parse(json)
-    p data
     parse_pageviews(data)
   end
   
   def cocktails_pageviews name, start, endd
-    dst = Config::HT_STAT_DIR + "/" + name + ".json"
+    dst = Config::REPORTER_DIR + "/" + name + ".json"
     
     # do not re-calculate stats older than four days
     if @last_updated - endd > 4 * DAY and File.exists?(dst)
@@ -199,6 +171,93 @@ class Analytics
   def update
     update_ratings
     update_reporter
+    update_stats
+  end
+  
+  def update_stats
+    endd = Time.now - DAY * 2
+    
+    
+    r = report({"dimensions" => "ga:date", "metrics" => "ga:visits,ga:pageviews"}, endd - DAY * 90, endd, 90)
+    # puts r
+    r = JSON.parse(r)
+    
+    total = r["totalsForAllResults"]
+    
+    stats = r["rows"]
+    stats.each do |e|
+      e[0] = Date.strptime(e[0], "%Y%m%d").to_time.to_i
+      e[1] = e[1].to_i
+      e[2] = e[2].to_i
+    end
+    stats.push({"total" => {"visits" => total["ga:visits"].to_i, "pageviews" => total["ga:pageviews"].to_i}})
+    
+    File.write(Config::VISITS_JSON, JSON.stringify(stats))
+    
+    
+    
+    r = report({"dimensions" => "ga:region", "metrics" => "ga:visits", "sort" => "-ga:visits"}, endd - DAY * 90, endd, 50)
+    # puts r
+    r = JSON.parse(r)
+    
+    total = r["totalsForAllResults"]
+    
+    stats = r["rows"]
+    stats.each do |e|
+      e[1] = e[1].to_i
+    end
+    stats.push({"total" => {"visits" => total["ga:visits"].to_i}})
+    
+    File.write(Config::CITIES_JSON, JSON.stringify(stats))
+    
+    
+    
+    r = report({"dimensions" => "ga:browser,ga:browserVersion", "metrics" => "ga:visits", "sort" => "-ga:visits"}, endd - DAY * 30, endd, 1000)
+    # puts r
+    r = JSON.parse(r)
+    
+    total = r["totalsForAllResults"]
+    
+    stats = r["rows"]
+    stats.each do |e|
+      e[2] = e[2].to_i
+    end
+    stats.push({"total" => {"visits" => total["ga:visits"].to_i}})
+    
+    File.write(Config::BROWSERS_JSON, JSON.stringify(stats))
+    
+    
+    
+    r = report({"dimensions" => "ga:browser", "metrics" => "ga:visits", "sort" => "-ga:visits"}, endd - DAY * 30, endd, 100)
+    # puts r
+    r = JSON.parse(r)
+    
+    total = r["totalsForAllResults"]
+    
+    stats = r["rows"]
+    stats.each do |e|
+      e[1] = e[1].to_i
+    end
+    stats.push({"total" => {"visits" => total["ga:visits"].to_i}})
+    
+    File.write(Config::BROWSERSP_JSON, JSON.stringify(stats))
+    
+    
+    r = report({"dimensions" => "ga:eventLabel,ga:browser,ga:browserVersion", "metrics" => "ga:uniqueEvents,ga:eventValue", "filters" => "ga:eventAction==error", "sort" => "-ga:uniqueEvents"}, endd - DAY * 30, endd, 500)
+    # puts r
+    r = JSON.parse(r)
+    
+    total = r["totalsForAllResults"]
+    
+    stats = r["rows"]
+    stats.each do |e|
+      e[3] = e[3].to_i
+      e[4] = e[4].to_i
+    end
+    stats.push({"total" => {"uniqueEvents" => total["ga:uniqueEvents"].to_i}})
+    
+    File.write(Config::ERRORS_JSON, JSON.stringify(stats))
+    
   end
   
   def update_ratings
@@ -250,7 +309,7 @@ class Analytics
     File.write(Config::HT_RATING_JSON, JSON.stringify(res))
   end
   
-  def get_month_borders year, month
+  def calc_month_borders year, month
     start = Time.new(year, month, 1)
     # jump to the next month (maybe year too)
     endd = start + 33 * DAY
@@ -282,7 +341,7 @@ class Analytics
       
       say "обновляю период «#{name}»"
       indent do
-        cocktails_pageviews(name, *get_month_borders(cur.year, cur.month))
+        cocktails_pageviews(name, *calc_month_borders(cur.year, cur.month))
       end
       @all << name
     end
